+++ meld/system/__init__.py
@@ -15,13 +15,17 @@ from meld.system.system import GeometricTemperatureScaler, REST2Scaler, RunOptio
 from meld.system.state import SystemState
 
 
-def get_runner(system, options, comm):
+def get_runner(system, options, comm, platform=None):
     if options.runner == "openmm":
         import meld.system
         import meld.system.openmm_runner
-        return meld.system.openmm_runner.runner.OpenMMRunner(system, options, comm)
+
+        return meld.system.openmm_runner.runner.OpenMMRunner(
+            system, options, communicator=comm, platform=platform
+        )
     elif options.runner == "fake_runner":
         from meld.system.runner import FakeSystemRunner
+
         return FakeSystemRunner(system, options, comm)
     else:
         raise RuntimeError(f"Unknown type of runner: {options.runner}")